Added data extraction from SQL and Redis and compared

+++ src/main/java/org/example/repository/FoundationRepository.java
@@ -17,7 +17,7 @@ public class FoundationRepository<T> {
         session.remove(entity);
     }
 
-    public T findById(Session session, Long id) {
+    public T findById(Session session, Integer id) {
         return session.get(entityClass, id);
     }
